requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-scheduler==0.10.1
filelock==3.16.1
icalendar==6.0.1
idna==3.10
lxml==5.3.0
oauthlib==3.2.2
PySocks==1.7.1
pyTelegramBotAPI==4.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==1.3.1
schedule==1.2.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
telebot==0.0.5
tqdm==4.66.5
tweepy==4.14.0
tzdata==2024.2
urllib3==2.2.3
